pyproject.toml
[project]
name = "bookclaw-langchain-py"
version = "0.1.0"
description = "Intelligent Agent system with MCP tool integration and subagent architecture"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # LangChain
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.3.0",
    # MCP
    "mcp[cli]>=1.0.0",
    # Web
    "starlette>=0.37.0",
    "uvicorn>=0.29.0",
    "httpx>=0.27.0",
    "fastapi>=0.135.2",
    # OpenAI compatible
    "openai>=2.30.0",
    # Utils
    "python-dotenv>=1.2.2",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
    "aiohttp>=3.13.5",
    # RAG
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "pymilvus>=2.6.11",
    "milvus-lite>=2.5.1",
    "hanlp>=2.1.3",
]
